feat: add pagination and Swagger configuration, update task management endpoints, and implement task creation and update DTOs

+++ src/main/java/com/vicente/taskmanager/scheduler/TaskScheduler.java
@@ -16,8 +16,10 @@ public class TaskScheduler {
     @PostConstruct
     @Scheduled(cron = "${spring.task.scheduling.cron}")
     public void execute() {
-        taskSchedulerService.updateOverdueTasks();
-        taskSchedulerService.deleteCancelledTasksOlderThan90Days();
-        taskSchedulerService.deleteDoneTasksOlderThan180Days();
+        if(taskSchedulerService.hasTasks()) {
+            taskSchedulerService.updateOverdueTasks();
+            taskSchedulerService.deleteCancelledTasksOlderThan90Days();
+            taskSchedulerService.deleteDoneTasksOlderThan180Days();
+        }
     }
 }
